feat(moac): update dependencies etc
The dependency update fixes a couple of sec vulnerabilities.

We switched back to official k8s client as all our fixes have been
merged to the upstream.

dirty-chai was removed because it's not possible to write TS tests
(definitions are different than expected). Instead we ignore the
eslint error which was why we started to use dirty-chai originally.