fix(nexus): handle IO errors per channel
IO errors are now handled differently such that errors on a channel, are
handled on a per channel basis.

When we are in submission, we simply remove the device from the
channel(s) and retry the IO by resubmitting it into the nexus.

Completion IOs are -- when completed, also retried. For each failed
device, we kick off one retire request which will pause the nexus while
in it does so. This prevents any new incoming IO.