feat(persistence): add persistent store
Add a persistent store which uses etcd as its backing store.

Each nexus saves a "NexusInfo" structure to etcd. The key for this
structure is the nexus UUID. The NexusInfo data includes the states of
the children and a "clean shutdown" flag.

The NexusInfo data structure has been defined in a new persistence.proto
file. This is necessary as the data structure is common between Mayastor
and MOAC.

There are 3 places where the persistent information is updated:
1. During nexus creation
2. When retiring a child due to an I/O error
3. During nexus destruction

An additional "-p" argument has been added to Mayastor which allows the
endpoint of the persistent store to be provided. If a port number is
omitted the default port number (2379) is used.

The "list nexus" RPC call has been modified such that it will not return
nexuses that are in the "Init" state. This prevents MOAC from publishing
a nexus before its state has been persisted (the nexus is set online
only after its NexusInfo has been saved to the persistent store).