feat: add persistent information interface to moac
Add a new interface that allows MOAC to determine which replicas can be used in the
creation of a nexus by using the persistent store (etcd).

On first creation all replicas are allowed to be used.
On recreations, only healthy replicas can be used.
If the last shutdown of the nexus was unclean, then only 1 replicas is choosen and for
the moment we choose a local replica, if possible.
The mayastor changes are being worked on separately.

On deletion of the volume we delete the entry from the persistent store.