test: add tests for the persistent store at the MOAC level
Add a persistent store with a mocked backend to the existing volume tests.

Add a new persistence test which does basic testing using a real etcd
server and uses a mock etcd client to test the various corner cases when
trying to create and recreate a nexus.

Configure mocha to use the typescript extension and replca the existing
javascript test index with a typescript file.
Add etcd to the nix-shell.