chore: revert persistent proto file
Add persistent nexus information type definition directly to MOAC.
This must match what mayastor writes to ETCD so any changes will
have to be made in tandem.
Note that MOAC does not test against a real mayastor instance so any
changes on mayastor will NOT trigger a failure of the moac tests.
Also replace snake_case with camelCase because the semistandard is not
running against TS files at the moment.