fix: don't succeed volume creation if it failed
When the volume creation fails we attempt to delete the volume. Before
this is complete we leave the volume in limbo and another create could
be tricked into an update, thus completing the CSI call with success.
To avoid this, check that the volume is not pending, unknown or
destroyed before allowing a volume update.

There is still an issue with this logic: if the destroy fails during
creation then nothing else will attempt the destroy again, leaving us
with a destroyed volume in the list, meaning the volume will get
stuck...