fix(mayastor): register callbacks after open
Properly closing a nexus child requires an event listener to be
registered in order to handle the resulting device removal event.

There is a bug in Nexus::try_open_children(), in which a nexus child may
be opened and then closed again before any event listener has been
registered, causing the close() to hang.

The fix is to ensure that the event listener is registered as soon as a
child is successfully opened