fix: don't add the same loopback device to the nexus
Other device types return an error when we attempt to open them twice.
The loopback's are special cases, because they're already open, eg: we
create a replica and then open the loopback device.
When a loopback device is reopened with a different UUID it ends up
overwriting the existing UUID and returning OK which sounds very bad.
Fix: Fail the open request if the UUID is not the expected one.

Related issue, because opening an existing loopback succeeds it means
that we can add nexus children to the nexus twice.
Fix: Fails NexusAddChild if the child is already part of the nexus!