fix(moac): avoid duplicating watcher upon restart
It seems to be an undocumented behaviour of the watcher library that
we are using, to call error callback even if there is no error upon
the first start. The solution is to register error callback after the
watcher is started.

I also made some changes to avoid printing error messages to the log
when the watcher restart was actually intentional.

Resolves: CAS-938