test(python): add a nexus create/destroy test and update FioSpdk
Fix docker-compose config so that container ms3 starts on a system
with only 8 cores.

Add a new test that creates 2 replicas shared over nvmf, 1 as
loopback. Create a nexus with all 3 replicas, then destroy the nexus,
for 10 repetitions.

Update the FioSpdk class to match the interface of Fio so that it
supports multiple filenames and add copies of existing Fio tests that
use FioSpdk instead, which removes the requirement for another VM.