feat(nexus): serialize gRPC requests
gRPC requests are received on the tokio runtime and dispatched to the
reactors by sending it a message. This made us run requests
in parallel which is what we want to avoid.

Additionally, the embedded json-rpc server only checks for incoming
messages every N ms. This change also makes the adopt that model by
holding the lock while it sleeps.

The lock implemented is an RW lock so we can make it more relaxed later
if we need to.