fix(test): io_timeout_reset fails too early
Depending on timing, if the KeepAlive + timeout is triggered before we've submitted the read IO,
the controller could be failed and the IO will fail before getting submitted.

This is likely because we now don't reset and retry to connect and simply fail the controller.

Fix: Increase the timeout count to avoid failing "too early" which fails the test.