feat(moac): decouple moac from mayastor
It is a preparation step for moving moac to a separate repository.
Major update to the documentation.

This commit solves two major problems:
1. Sharing proto files between mayastor and moac
2. Removing dependency on nix and using docker to build moac image

add 1. We use curl to download proto files from the mayastor repo.
It does not make sense now but will in the future. Everytime a change
to proto files in mayastor repo is made, the commit hash in a script
needs to be changed to pull these new changes. It is a conscious
decision resulting into a deterministic build of moac.

add 2. We still provide nix-shell for setting up dev environment for
moac, however that's all what we use the nix for. The script
csi/moac/scripts/release.sh is a simplified version of its
mayastor counterpart.

Resolves: CAS-958