fix(nix): revert back to h2 fork
Unfortunately, we still need to carry one patch around. To keep with the
trend of not having to update cargoSha256 with every cargo update we now
only need to specify the sha256 of the crates we patch. Note that this
will only work if the follow what the nix implementation expects, which
is a tag (rev).