feat(nexus): change create_v2 gRPC to have separate name and uuid
Add a separate name field to the CreateNexusV2 gRPC, which is a
freeform string used directly in the Nexus structure in Rust instead
of being derived from the uuid field by prefixing "nexus-" to it. This
is in turn turn used as part of the subnqn when the nexus is
published.

Change nexus_lookup() to use the nexus name directly if it is not a
uuid to ensure that nexuses created with v1 and v2 gRPCs can be looked
up.

This in turn requires a ListNexusV2 gRPC that returns both name and
uuid fields. Update the Python test to use both new calls. Revert part
of the test_nexus mocha test back to the v1 CreateNexus to avoid too
much churn throughout that test, plus it will be superceded by the
Python test anyway.