chore(build): rust-overlay and container image
This removes the test binaries from the container image as they are not
needed. Also add /bin to PATH and wrap mayastor-client to mctl.

Lastly, the mozilla-overlay isn't very flexible to use. Switcht to a
more flexible overlay which, for example, does not take for ever to copy
HTML files when building images.