feat: build against SPDK 21.07
Add pyelftools as a build input as the DPDK build requires it.

Upstream has made struct spdk_thread private so switch to public APIs
to obtain thread ID and name. The spdk_io_channel struct has also been
made private, use the public API for IO channel name but include the
private header as there is no other way of obtaining the name of the
thread in the IO channel.

Sync with upstream changes to structs spdk_bdev_alias,
spdk_bdev_nvme_opts and spdk_iscsi_opts. There is now a separate
timeout for NVMe admin commands, which defaults to the same 5s as IO
commands and is configurable with the NVME_TIMEOUT_ADMIN_US
environment variable.