fix(bdev): delete alias when destroying bdev
A test in replica_test.js creates and destroys the same aio bdev
twice, which in SPDK 21.07 leads to a panic in lvs_pool.rs when
attempting to unwrap Bdev::bdev_uri(). The second time the aio bdev
is created, adding the alias fails with EEXISTS but this is usually
ignored.

This is due to the new red-black tree implementation for bdev aliases
which results in the alias persisting in the red-black tree after the
bdev was deleted without deleting its alias first.

Fix this by introducing Share::bdev_uri_original() which returns the
URI used to construct the bdev without adding the uuid query. Use it
in lvs_pool when destroying the bdev to ensure that the bdev alias
is removed.