chore(ci): don't lint PRs
While it'd be great to lint the PR title and description, the commitlint
isn't really fitting for this purpose. The 74 character limit makes the
PR description looks really strange on today's high-res displays and
also makes it very annoying to write the descriptions (since browsers
don't show the cursor position, people end up writing the description in
their editor first and then pasting it in the browser). This limit also
makes it hard to add links.

Using commitlint also assumes 1 commit per PR and we want to encourage
atomic commits, which often implies multiple commits per PR.

So until someone comes up with a better way to link PRs, let's not do
that.