fix(nvmx): handle failures from processing completions for admin commands
Handle it by scheduling a future once, and only once, to destroy the
device if spdk_nvme_ctrlr_process_admin_completions() returns an
error, usually -ENXIO if the remote child is unreachable when
handling admin commands such as keep-alive.

This means an unreachable remote child on an idle nexus is now noticed
and the child is marked as degraded. In the case where the nexus is
under IO load and 2 remote replicas are unreachable at almost the same
time, both replicas are now consistently degraded.

The persist_io_failure cargo test reveals that the child now ends up
in a degraded state but the etcd entry still reveals it to be healthy
so update its entry when processing the device removal event. This
means the children are marked as "unhealthy" after the nexus is shut
down as they are closed so change the persist_clean_shutdown test
accordingly.

Fix the existing io_timeout_reset cargo test to now expect
device_destroy to return an error as the admin command completion
handler will have already destroyed it. In the nexus_multipath tests,
add a 1s delay after nvme connect as the nvme list immediately
following more frequently returns no Mayastor devices.