fix(grpc-test): allow time for namespace removal to be actioned
In the destructive nvmf test, reinstate the call to get the list of
bdevs after removing the namespace from the nvmf subsystem, which
was skipped after the switch to nvmx as it was no longer relevant.
This allows time for the namespace removal to be processed and for
the nexus and child to end up in the correct states of fauled and
degraded, respectively.

This improves the reliability of this test, which can fail on rare
occasion.