fix(nvmx): ensure child is faulted after handling admin command failure
Change the handling of errors from
spdk_nvme_ctrlr_process_admin_completions() to sending an
AdminCommandCompletionFailed event, which in turn causes the listener
to mark the child as faulted due to IO errors, then retire child, much
as an IO completion failure would. This ensures that a child that is
inaccessible ends up in the state faulted due to IO error, regardless
of whether the error is first noticed in the IO or admin completion
paths.

Revert the changes that persist the child state when it is removed for
reasons other than faulted due to IO error, the corresponding changes
to the persist_io_failure and persist_clean_shutdown cargo tests, and
the io_timeout_reset cargo test.

Change the pytests that test killing a container that hosts the child
to expect the child to be in faulted state.