chore(reactor): drop Future impl for `&'static Reactor`
This impl does strange (working on master reactor even though
implemented for the generic type) and bad (does a blocking timeout)
things. Moreover, it's not really needed since everything it does is
done already by the spdk poller.

This commit breaks the grpc code though as tonic assumes to be run in
the tokio runtime and currently `Reactor::block_on` doesn't use tokio's
runtime. This will be fixed in a following commit.