fix(nexus): nexus_bdev::io_is_supported return value
Fix nexus_bdev::io_is_supported() to return False if any child does
not support the given IoType, as the comment suggests.

This was exposed when sending a write-zero to a nexus comprising a
local and remote replica where local replica correctly wrote zeroes
but the remote replica did not. The local replica uses the lvol bdev
directly, which supports write-zero, whereas the remote replica uses
nvmx, which does not. The latter translates the write-zero into a
deallocate (NVMe for unmap) which ends up being ultimately ignored
by the aio bdev, via the lvol bdev, on the remote replica. Fixing the
function causes the bdev layer to emulate write-zero on an unsupported
bdev, such as nvmx, by sending a write of a block of 0s.

Add a cargo test to test the above scenario and rename the existing
nexus_multipath.rs to nexus_io.rs to better reflect its test area.