feat(nvmx): implement write_zeroes
Implement write_zeroes for nvmx instead of converting it to unmap
and mark it as a supported IoType. This allows the write-zeroes op
to be sent to the remote replica as a more efficient Write Zeroes
command instead of a Write command with a block of 0s.

At the remote replica, the lvol bdev handles write-zeroes correctly
regardless of whether the underlying bdev supports it or not.