fix(nexus): persist state of newly added child to etcd
Introduce PersistOp::AddChild to add the state of a newly added child
to NexusInfo and persist it to etcd. Update the state of the new
child when it has rebuilt successfully. Mark a removed child as
unhealthy so that a recreated nexus will not use that child.

Augment the existing persist_io_failure cargo test to add a new child
and check its persisted state before and after rebuild has completed,
and after it has been removed.