fix(nvmx): support for Unmap and WriteZeroes IoTypes
Make support for these IoTypes dependent on the controller's ONCS
(Optional NVM Command Support) field instead of hardcoding. This
ensures that Write Zeroes is not inadvertently supported by nvmx if
the NVMe target does not advertise support. This now makes the NVMf
target report support for Dataset Management as the lvol bdev does
support that, regardless of whether the underlying bdev supports it.

Verify in a pytest that the NVMf target for a published nexus with
remote replicas supports these operations as they essentially reflect
what is supported by the lvol bdev at the remote end.