fix(csi): node_unpublish_volume() fails if nvmf target is unreachable
When performing I/O on a volume has failed due to the underlying nvmf
target being unreachable, subsequent attempts to unpublish the volume are
ineffective, resulting in an endless loop where node_unpublish_volume()
is called repeatedly with no effect.

Cause:
When the underlying nvmf target has timed out, and is thus disconnected,
we are left with a mount to a device that no longer exists. In the
case of a filesystem mount, any attempt to access the corresponding
mount_point fails with an EIO error, although an umount() would succeed
at this point. At present, node_unpublish_volume() implicitly performs a
stat() on the mount_point to determine what type of bind mount is present
(either a block special file or directory). It then proceeds to perform
an umount(), followed by deletion of the file or directly that is the
mountpoint. If the underyling nvmf device has been disconnected, then the
initial stat() can fail with an EIO. Unfortunately, this is incorrectly
interpreted to mean hat the mountpoint does not exist, and that the
device has already been unmounted. Thus node_unpublish_volume() assumes
that there is nothing to do and returns (reporting success). The result
is repeated calls to node_unpublish_volume() that have no material effect.

Fix:
The fix is to change the order in which the stat() and umount() are
performed. The required unmount operation is identical in this case,
regardless as to whether the mount_point is a file or directory. The
umount() will succeed even if the nvme device no longer exists and,
assuming the umount() succeeds, a subsequent stat() on the mount_point
no longer fails, thus allowing the deletion of the directory or file
as appropriate.

resolves CAS-817