fix(nexus_v2): add uuid field to the nexus
The CreateV2 was changed to use uuid's but all the lookups are made using
the name, even though the arg is called uuid...
This should be properly addressed as part of 1107 but for now use the
workaround of looking up by uuid if by name fails.

When a nexus is published the NQN is set to the name but the namespace uuid's are
set to the nexus UUID. This breaks the CSI plugin logic, as it expects them to be
the same...
As a WA for the moment, when using the V2 api set the bdev uuid to match
the name, if the name is a UUID. The uuid arg is then used as a means to identify
the nexus only.
This means the nqn and the namespace uuid is now using the nexus name, which in
case of the control plane on k8s, is the volume uuid.

Also tidiy up some corner cases with UUID's which may be reused with
different names - spdk does not seem to treat UUID as unique...

With V2 there is no "nexus-" prefix to add env variable to the CSI plugin
so that we may keep the same behaviour for MOAC testing.