feat(bench): add benchmarks crate for io-engine operations
Adds nexus creation benchmark via gRPC and directly.
The benchmarks are added on a different crate (io-engine-bench) because otherwise on every change
on the bench code every io-engine crate binaries has to be built which takes time.
The downside is that io-engine binaries release or debug need to be built prior to running the
benches.
(This can easily be changed if we want to)

The benchmark needs to run from the io-engine-bench folder because:
1- we have a runner to run it as root (required by the in-binary benchmark)
2- we have code in the runner to chown the results to $USER and copy them back and forth
from the io-engine-bench/results/criterion location and the target location.

To run the benchmark in release:
RUST_LOG=disable cargo bench -p io-engine-bench

To run the debug:
RUST_LOG=disable $RUST_NIGHTLY_PATH/bin/cargo bench -p io-engine-bench --profile=dev

The criterion results are placed in the cargo target folder. You may direct your browser to this
location to inspect the results.