fix(nexus): fixing loopback device destruction
If nexus creation fails and one of the children is a loopback pseudo-device,
child close operation never completes: when a loopback pseudo-device
is destroyed, SPDK bdev remove event is not sent because loopback is not
a real bdev. Instead, loopback pseudo-device tried to unplug itself
from nexus, but nexus cannot be found because it failed to register.

Signed-off-by: Dmitry Savitskiy <dmitry.savitsky@datacore.com>
Signed-off-by: Dmitry Savitskiy <dmitry.savitskiy@datacore.com>