fix(test): don't print the device name
The Debug for NexusChildren is formatting the string with the device name, which can cause
a segfault after the device has been destroyed.
The device is an Option but it seems it's not taken at the point of destruction... this
might point out there's some logic issue but since this whole section of code seems quite
complicated it's better for now to simply remove the device name from Debug.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>