fix(thin-provisioning): don't leak replicas when wipe_super fails
When the clear method of a pool is not unmap, we wipe the first 8M of a replica after creation.
When using thin provisioning this may fail with ENOSPC, and in this case clean up the created
replica as if we can't even write this much, it's probably not worth using it.

How should we handle this in the control-plane? For now we should probably add a slack of
8M when determining if a pool is usable.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>