feat(fio-spdk): add userspace fio with spdk engine
Adds a new fio test image that can use the spdk ioengine to send IOs to a remote nvmf target.
Also when not part of CI restricts the docker image tag to the alias; this is helpful when
building locally to avoid collecting a bunch of images with hashes if we only care
about a specific tag, eg: develop.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>