fix(nvmf): don't enable ANA reporting for lvol replicas
We only need ANA reporting on the nexus and having the replicas ANA seems to cause some
issues when the nexus connects to them.
Adding a new `ShareProps` which can be used for various share properties - also added a
allowed host option which for the moment defaults to any that will likely be helpful
to gate access to the targets.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>