fix(nvme-reservation): acquire without preempt
When stealing a reservation issue a second acquire is necessary as SPDK is not always updating the
namespace reservation.
When registering a new key with Register Action REPLACE and Ignoring Existing Key, the
registration  succeeds and the key is replaced but the registration is not changed in
the namespace. In this case the report contains the wrong key as the holder so the
previous acquire is not sufficient.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>