feat(nvme-resv): add nvme reservation persistence through powerloss
Enable Persistence Through Power Loss which is setup with SPDK when creating a namespace.
The base directory is specified via a cli argument, eg: /host/tmp/ptpl
Each resource gets its own entry within the base, eg:

├── nexus
│   ├── e0be2258-f2ac-4691-a607-a91e7145abe4.json
└── pool
    ├── 0b1c02c1-2840-469b-b101-105b0120f663
    │   └── replica
    │       └── 50a1e576-d475-4f66-9929-12ddbabbd0a8.json
    ├── 14d860af-4e8b-48e8-9328-88672f0d6fd0
    │   └── replica
    │       └── 207b1b37-df8a-4a24-9f4a-54a2988cc884.json
    └── 177197c0-c2dc-4a34-bdf7-1d3423e7f35e
        └── replica
            └── 7f49aca4-70ed-4a68-9d10-c8269bdcf98a.json

To achieve this we've added a new trait `PtplFileOps` which any resource can implement.
The `Share` trait is then enhanced to include a new `PtplProps` which includes the
full path for the ptpl json file.

When a resource is deleted, we have to delete the file as well.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>