refactor: remove unsafe constructs from rebuild
The previous rebuild code was making some assumptions about the way the rebuild job runs.
The code is ever run in the master reactor, so there can never be two tasks concurrently
touching the rebuild job - however with time seems in certain cases between awaits the
rebuild job may be reused which is likely bad and could explain some weird issues
we've seen with rebuild as of late.

This change removes the unsoundness of the rebuild at the cost of a few locks and
channels. However these probably won't affect the rebuild performance much as they
will be largely uncontended.
The rebuild job is kind of split into frontend and backend. The backend is where
the rebuild actually happens and we have a channel used to communicate between
frontend and backend.

This change also allows to, in the future, have dedicated cores just for rebuilds for example.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>