fix(rebuild): set final rebuild stats
When rebuild completes, right before the nexus completes the rebuild callback it's possible
that we get a stats request and in such case we have no stats to give.
To address this, then the rebuild backend completes it now sets the final rebuild stats in
the states, which will allow the stats to be retrieved.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>