fix(lvs/replica): destroy using pool identifier
Adding the pool to the request allows us determine whether a pool is loaded or not.
Example: if a pool is not loaded then we cannot destroy the replica and we must return an
adequate error code back as it's the control-plane job to import pools before attempting
to destroy replicas.

To distinguish from previous not found implementations, a new header with the gtm ticket
number if added.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>