fix(state/degraded): control-plane does nothing for degraded
At the moment degraded means for control-plane: do nothing.
It's worse though... if it's enospc, then we have special behaviour.. but for partial
rebuild we have special behaviour for faulted...
We need to fixup our behaviour, but for now let's keep things as they were, where degraded
only applies for transient config states, rebuilding and enospc.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>