fix(fs/shutdown): pause nexus when faulted
When the last healthy child of a nexus is faulted, errors are propagated up the stack causing the
filesystem to shutdown itself down.

Instead I propose that we simply leave the nexus in a paused state, allowing the initiator to
start the re-connection cycle, giving us some “time” to recreate the nexus again when the
replica is available again.

todo: we might want to add a time-limit after which we resort to failing IOs?

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>