fix(nexus/io): set nvmf subsystem as frozen
Multiple children may be retired at "the same time". In case the first retires "wins" and
pauses the subsystem, if we simply leave it paused at the end, the other retires will remain
stuck in the pause.
Instead, introduce a new state of Frozen, which allows other pauses to execute their paused
code and at the end the subsystem remains frozen.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>