feat(wipe/replica): add rpc to a wipe a replica using methods
Adds rpc wipe replica using streaming on the server side which allows the client to
receive updates as the wipe is ongoing. This is mainly useful for large replicas
which may take a very long time to wipe!

We provide a few wipe methods, none, zero, unmap and pattern but only none and zero
are implement for now.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>