fix(nvmf): removing namespace before destroying subsystem
When nexus Bdev is destroyed, SPDK sends out SPDK_BDEV_EVENT_REMOVE event.
In response to this event, SPDK NVMF subsystem removes the corresponding
namespace.
When io-engine destroys the subsystem, it first stops it.
Sometimes, the event for destroying nexus Bdev arrives before the subsystem is
destroyed but it is stopped, which leads to an assertion
in SPDK (actual_old_state == expected_old_state).

Signed-off-by: Dmitry Savitskiy <dmitry.savitskiy@datacore.com>