fix(pcie): spdk requires nvme path id
When destroying a pcie nvme bdev we must not specify the nvme path id.
Also it was noticed that when destroy fails we end up with no alias which
could cause some panics. For this particular case we now add the alias
back if the bdev is still present though we need a more robust fix..

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>