fix(iobuf): fixup large iobuf pool size
Addresses warning from spdk itself:
mayastor::spdk:transport.c:282]
The num_shared_buffers value (2048) is larger than the available iobuf pool size (1024).
 Please increase the iobuf pool sizes.

Also reduce size to power of 2 - 1 as recommended in spdk ticket:
It’s always best to specify a value for –n that is one less than a power of two (2^x -1)
because of the way rings work in DPDK. You end up allocating a ring twice the size for
the one extra element between 511 and 512 for example.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>